+++ src/components/ListTemplate/listTemplate.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import {
-  FormattedMessage,
-  FormattedTime,
-} from 'react-intl';
+import { FormattedMessage } from 'react-intl';
+
+import { FormattedTime } from '@folio/stripes/components';
 
 import {
   CheckboxColumn,