+++ src/components/MatchCriterion/view/IncomingSectionStatic/StaticValueDateRange.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  FormattedDate,
-  FormattedMessage,
-} from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
 import {
   Layout,
   Headline,
   NoValue,
+  FormattedDate,
 } from '@folio/stripes/components';
 
 import css from '../ViewMatchCriterion.css';